requirements.txt
# Algorithmic Trading Platform Requirements

## Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.1
httpx==0.25.2
celery==5.3.4
apscheduler==3.10.4

## Trading & Market Data
fyers-apiv3==2.0.0
pandas==2.1.4
numpy==1.25.2
ta-lib==0.4.28

## Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.34.0

## Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0
structlog==23.2.0

## Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

## Utilities
python-dotenv==1.0.0
tenacity==8.2.3
aiofiles==23.2.1
jinja2==3.1.2
